requirements.txt
# Core ML
torch>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Training
unsloth>=2024.1
transformers>=4.36.0
datasets>=2.14.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
trl>=0.7.0
peft>=0.7.0

# API & Async
aiohttp>=3.9.0
openai>=1.0.0
tiktoken>=0.5.0

# Monitoring
wandb>=0.16.0

# Utilities
huggingface_hub>=0.19.0
scikit-learn